First release
This release is the first complete release where the usability and error checking has been greatly improved. The configuration node has the ability to test the Hydrawise controller on the backend.
The command node now provides easy access to the available zones.
The info command now lists all available info to be listed to a debug node.